apps/scraper/requirements.txt
psycopg2-binary
requests
beautifulsoup4
python-dotenv
apscheduler
fastapi
uvicorn
